requirements.txt
beautifulsoup4==4.12.3
click==8.1.7
loguru==0.7.2
lxml==5.2.2
pydantic==2.8.2
python-dotenv==1.0.1
requests==2.32.3
selenium==4.23.1
